Update compatible OS name and versions
Previous to this commit, the metadata.json had the supported OS name as
"nexus" and the only supported version as 7.0.  Since the metadata.json
file was incorrect, the module's Forge page was also incorrect. At the
request of Cisco, the name should be NX-OS. Scott Garmen reports the
compatible versions are 7.02, 7.021, and 7.03.  This commit updates the
metadata.json file with the correct information