pim beaker tests
Also had to update the existing scripts to clean up stale configs
from other tests.

All of the routedintf tests now pass, except for an issue with the
negative test for channel-group, which correctly hits an error but
for which beaker does not get exit code [1], so I'm temporarily
removing that test.