remove test_title_patterns
It was useful but it adds another level of complexity to the tests so I just removed it and converted the title pattern tests to use the regular tests hash and the new test_harness_run() method. One hash, one harness call.

The rest of the changes address the remaining code review comments for PR #176