(FM-7454) conversion of search_domain to agentless
search_domain is a netdev_stdlib type which is deprecated, in order to keep feature parity the provider has still been converted.

`search_domain` makes use of `Cisco::DomainName` where it indicates the potential for more than one, however this is not the case and only one `domainname` will ever be present on the device.